LayoutTests/crypto/subtle/aes-gcm-import-key-unwrap-ec-raw-key-expected.txt

Test unwrapping an EC raw key with AES-GCM using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(unwrappedKey) is expectedRawKey
PASS successfullyParsed is true

TEST COMPLETE
